requirements.txt
accelerate==1.6.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
bitsandbytes==0.42.0
cattrs==25.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.2
coremltools==8.2
cycler==0.12.1
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.3.0
fastapi==0.117.1
ffmpy==0.6.1
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
gradio==5.11.0
gradio_client==1.5.3
h11==0.16.0
hdbscan==0.8.39
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
latex2mathml==3.78.1
# Editable install with no version control (llava==1.2.2.post1)
# -e /Users/riddhiman.rana/Desktop/Coding/orion-research/ml-fastvlm
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==2.1.5
matplotlib==3.10.6
mdurl==0.1.2
mpmath==1.3.0
neo4j==5.26.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.7.0.72
orjson==3.11.3
packaging==25.0
pandas==2.3.3
peft==0.13.2
pillow==11.3.0
polars==1.34.0
polars-runtime-32==1.34.0
protobuf==6.32.1
psutil==7.1.0
py-cpuinfo==9.0.0
pyaml==25.7.0
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
rich==14.1.0
ruff==0.13.1
safehttpx==0.1.6
safetensors==0.6.2
scikit-learn==1.2.2
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==3.3.1
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
svgwrite==1.4.3
sympy==1.13.1
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.21.0
tomlkit==0.13.3
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.3
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.204
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.37.0
wavedrom==2.0.3.post3
websockets==14.2
